# Conflicts:
#	src/main/java/com/construtech/buildsphere/platform/BuildspherePlatformApplication.java
#	src/main/java/com/construtech/buildsphere/platform/project_management/interfaces/rest/ProjectController.java
#	src/main/java/com/construtech/buildsphere/platform/shared/domain/model/aggregates/AuditableAbstractAggregateRoot.java
#	src/main/java/com/construtech/buildsphere/platform/shared/domain/model/entities/AuditableModel.java
#	src/main/java/com/construtech/buildsphere/platform/shared/infrastructure/documentation/openapi/configuration/OpenApiConfiguration.java
#	src/main/java/com/construtech/buildsphere/platform/shared/infrastructure/persistence/jpa/configuration/strategy/SnakeCaseWithPluralizedTablePhysicalNamingStrategy.java
#	src/main/resources/application.properties

+++ src/main/java/com/construtech/buildsphere/platform/resourceManagement/domain/model/valueobjects/Project.java
@@ -3,5 +3,5 @@ package com.construtech.buildsphere.platform.resourceManagement.domain.model.val
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record Project(int projectId) {
+public record Project(Long projectId) {
 }
